Split Map+Cartographer and Fenwick
